fix: Addressed all warnings about narrowing conversions in `core-forms`

+++ extra/appBasicForm.cpp
@@ -39,7 +39,7 @@ try
 	startupBanner();
 	
 	BasicForm mainDlg;
-	return mainDlg.showModal();
+	return static_cast<int>(mainDlg.showModal());
 }
 catch (std::exception const& e)
 {
